test(task-migrate-clean-batch): 迁 crystal_renderer 相机矩阵 2 用例到 gui_unit_test
420.5 Step 2。整文件迁移，纯相机矩阵数学，零 g_state 引用。
原 category 本就叫 "unit"，作者自证纯逻辑。

- 新增 test/unit-correctness/gui/test_crystal_renderer.cpp（2 个 TEST）
- 删除 test/gui/visual/test_gui_crystal_renderer.cpp + 注册声明/调用点
- 顺手删掉该文件里从未被引用的 constexpr kPi

核验：断言 IM_CHECK 7 → EXPECT_TRUE 7 一一对应；
gui_test 静态用例数 559→557；gui_unit_test 16→18 全绿。
