+++ test/gui/CMakeLists.txt
@@ -24,7 +24,6 @@ add_executable(gui_test
     visual/test_gui_visual.cpp
     visual/test_gui_render.cpp
     visual/test_gui_bg.cpp
-    visual/test_gui_crystal_renderer.cpp
     visual/test_gui_auto_ev.cpp
     visual/test_gui_capture_smoke.cpp
     visual/test_gui_lens_projection.cpp
