+++ test/unit-correctness/gui/test_crystal_renderer.cpp
@@ -1,0 +1,131 @@
+// Unit tests for CrystalRenderer::ComputeMvp's geometric contracts (camera
+// view-rotation split, world-coord interaction). Pure math — no GL context
+// required, which is why they live in gui_unit_test rather than gui_test.
+//
+// These tests guard:
+//   - V_rot = Rx(-kCameraTiltDeg) sign convention (camera elevated above origin
+//     looks DOWN at the crystal — world +z visible above screen center)
+//   - Y-Z swap equivalence: under the GUI mesh swap, mesh +y represents world
+//     +z, so the kColumn chain rotation (which maps c-axis to world +x in core
+//     conventions) — when applied to a c-axis-aligned mesh vertex — produces a
+//     screen-y near zero (c-axis horizontal).
+//
+// Cross-module guard: if BuildCrystalMeshData ever changes the swap direction,
+// the kColumn sub-test fails immediately because the mesh-frame chain output
+// no longer aligns with the V_rot convention.
+
+#include <gtest/gtest.h>
+
+#include <array>
+#include <cmath>
+#include <cstring>
+
+#include "gui/axis_presets.hpp"
+#include "gui/crystal_renderer.hpp"
+
+namespace {
+
+// Identity 4x4 column-major.
+void Identity4x4(float m[16]) {
+  std::memset(m, 0, 16 * sizeof(float));
+  m[0] = 1.0f;
+  m[5] = 1.0f;
+  m[10] = 1.0f;
+  m[15] = 1.0f;
+}
+
+// Multiply column-major 4x4 MVP by a homogeneous (x, y, z, 1) point. Returns
+// {clip_x, clip_y, clip_z, clip_w}; perspective divide is the caller's job.
+std::array<float, 4> Mvp4(const float mvp[16], float x, float y, float z) {
+  return {
+    mvp[0] * x + mvp[4] * y + mvp[8] * z + mvp[12],
+    mvp[1] * x + mvp[5] * y + mvp[9] * z + mvp[13],
+    mvp[2] * x + mvp[6] * y + mvp[10] * z + mvp[14],
+    mvp[3] * x + mvp[7] * y + mvp[11] * z + mvp[15],
+  };
+}
+
+}  // namespace
+
+// unit/crystal_renderer_camera_tilt_world_z_above_center — With identity
+// model rotation and zoom=1, a vertex at world +z (= mesh +y under the GUI
+// Y-Z swap) must project ABOVE the screen center AND its surface normal
+// must point toward the camera (not away). The "above center" assertion
+// alone is insufficient: with V_rot of either sign the top vertex appears
+// above center because both "top tilted toward camera" and "top tilted
+// away from camera" raise its screen y. The visible-face check (normal·
+// (+z_eye) > 0) is what pins the V_rot sign — required for kPlate to show
+// its top face (face 1) under the elevated-camera intent rather than the
+// bottom face (face 2).
+TEST(CrystalRenderer, crystal_renderer_camera_tilt_world_z_above_center) {
+  float model[16];
+  Identity4x4(model);
+  float mvp[16];
+  lumice::gui::CrystalRenderer::ComputeMvp(model, /*zoom=*/1.0f, /*width=*/512, /*height=*/512, mvp);
+
+  // World +z is mesh +y (post Y-Z swap). With camera elevated by
+  // +kCameraTiltDeg looking down, this vertex must land above center.
+  auto top = Mvp4(mvp, 0.0f, 1.0f, 0.0f);
+  EXPECT_TRUE(top[3] > 0.0f);
+  float ndc_y_top = top[1] / top[3];
+  EXPECT_TRUE(ndc_y_top > 0.0f);
+
+  // Symmetric sanity: world -z (mesh -y) must land below center.
+  auto bot = Mvp4(mvp, 0.0f, -1.0f, 0.0f);
+  EXPECT_TRUE(bot[3] > 0.0f);
+  float ndc_y_bot = bot[1] / bot[3];
+  EXPECT_TRUE(ndc_y_bot < 0.0f);
+
+  // **Sign-pinning assertion**: the top face's normal (world +z = mesh +y
+  // for identity model) must point TOWARD the camera in eye space, i.e.,
+  // its eye-space z component is positive (eye-space +z is out of screen
+  // toward the viewer in OpenGL). Below: transform the (0,1,0) direction
+  // by V_rot only — translation drops out for direction vectors, and
+  // model is identity. Equivalently we read the y column of (V_rot ·
+  // model) and check its z entry. This catches a sign flip that the
+  // ndc_y > 0 check above cannot.
+  // V_rot is the only rotation in the view chain (model = identity), so
+  // its 3x3 entries are recoverable by inverting the projection's effect
+  // on a direction vector — easier path: re-derive V_rot expectations
+  // from a known input. We compute the eye-space position of the top
+  // vertex and check that, after subtracting the camera-back translation,
+  // its z is positive (toward camera).
+  // dist = ComputeDist(zoom=1) ≈ 3.73; eye-space z of (0,1,0) post-V_rot
+  // is sin(kCameraTiltDeg) ≈ 0.26 (for positive Rx angle) or -0.26
+  // (negative angle). After T(0,0,-dist) the absolute z is ≈ -dist;
+  // adding back +dist recovers the rotation contribution.
+  auto top_world_eye = Mvp4(mvp, 0.0f, 1.0f, 0.0f);  // already in clip space
+  // Reconstruct eye-space z from clip: clip_w = -eye_z (OpenGL).
+  float eye_z_top = -top_world_eye[3];  // = -dist + V_rot's z contribution
+  // Compare against what dist alone would produce (no rotation): -dist.
+  // If V_rot tilts top TOWARD camera, eye_z_top is closer to zero
+  // (less negative); if AWAY, it is further from zero (more negative).
+  float dist_alone = -lumice::gui::CrystalRenderer::ComputeDist(1.0f);
+  EXPECT_TRUE(eye_z_top > dist_alone);  // strictly greater (less negative) → top tilts toward camera
+}
+
+// unit/crystal_renderer_kcolumn_caxis_horizontal — With the kColumn default
+// (chain + swap-wrapped to mesh frame) as the model matrix and a vertex at
+// the c-axis (mesh +y, which represents world +z under the swap), the rotation
+// maps mesh +y to mesh +x (horizontal). With V_rot only adding a small camera
+// elevation, the resulting screen y stays close to zero — the c-axis appears
+// horizontal. This guards the joint correctness of (a) ChainRotationToMatrix,
+// (b) the swap wrap inside DefaultPreviewRotation, and (c) the V_rot sign.
+TEST(CrystalRenderer, crystal_renderer_kcolumn_caxis_horizontal) {
+  // Use DefaultPreviewRotation — the same path used by the runtime, which
+  // applies the mesh swap wrap so the rotation is geometrically valid for
+  // mesh-space vertices.
+  float model[16];
+  lumice::gui::DefaultPreviewRotation(lumice::gui::AxisPreset::kColumn, nullptr, model);
+
+  float mvp[16];
+  lumice::gui::CrystalRenderer::ComputeMvp(model, /*zoom=*/1.0f, /*width=*/512, /*height=*/512, mvp);
+
+  // Apply MVP to mesh +y (the c-axis). After the wrapped chain the c-axis
+  // lies along mesh +x (horizontal); kCameraTiltDeg adds a tiny screen-y
+  // offset that stays well within tolerance.
+  auto p = Mvp4(mvp, 0.0f, 1.0f, 0.0f);
+  EXPECT_TRUE(p[3] > 0.0f);
+  float ndc_y = p[1] / p[3];
+  EXPECT_TRUE(std::fabs(ndc_y) < 0.1f);
+}
